+++ tests/test_power.py
@@ -1,12 +1,27 @@
 """Tests for the _ATEN_PE6108G PDU class"""
 
+import configparser
 from contextlib import contextmanager
 from unittest.mock import patch
 
 import pytest
 from pysnmp.proto.rfc1902 import Integer
 
-from _stbt.power import uri_to_power_outlet
+from _stbt.power import config_to_power_outlet
+
+CONFIG_INI = """
+[device_under_test]
+power_outlet = myaten
+
+[power_outlet myaten]
+type = aten
+address = mock.host.name
+outlet = 1
+"""
+
+
+CONFIG = configparser.ConfigParser()
+CONFIG.read_string(CONFIG_INI)
 
 
 def mock_data(int_value):
@@ -31,7 +46,7 @@ oid = "1.3.6.1.4.1.21317.1.3.2.2.2.2.{0}.0".format(outlet + 1)
 def test_aten_get_on():
     with mock_command_gen() as mock_command:
         mock_command.getCmd.return_value = mock_data(2)
-        aten = uri_to_power_outlet('aten:mock.host.name:1')
+        aten = config_to_power_outlet(CONFIG)
 
         result = aten.get()
 
@@ -41,7 +56,7 @@ def test_aten_get_on():
 def test_aten_get_off():
     with mock_command_gen() as mock_command:
         mock_command.getCmd.return_value = mock_data(1)
-        aten = uri_to_power_outlet('aten:mock.host.name:1')
+        aten = config_to_power_outlet(CONFIG)
 
         result = aten.get()
 
@@ -52,7 +67,7 @@ def test_aten_set_on():
     with mock_command_gen() as mock_command:
         mock_command.setCmd.return_value = mock_data(2)
         mock_command.getCmd.side_effect = [mock_data(n) for n in (1, 1, 1, 2)]
-        aten = uri_to_power_outlet('aten:mock.host.name:1')
+        aten = config_to_power_outlet(CONFIG)
 
         aten.set(True)
 
@@ -63,7 +78,7 @@ def test_aten_set_off():
     with mock_command_gen() as mock_command:
         mock_command.setCmd.return_value = mock_data(1)
         mock_command.getCmd.side_effect = [mock_data(n) for n in (2, 2, 1)]
-        aten = uri_to_power_outlet('aten:mock.host.name:1')
+        aten = config_to_power_outlet(CONFIG)
 
         aten.set(False)
 
@@ -74,7 +89,21 @@ def test_aten_set_timeout():
     with mock_command_gen() as mock_command:
         mock_command.setCmd.return_value = mock_data(1)
         mock_command.getCmd.return_value = mock_data(2)
-        aten = uri_to_power_outlet('aten:mock.host.name:1')
+        aten = config_to_power_outlet(CONFIG)
 
         with pytest.raises(RuntimeError):
             aten.set(False)
+
+
+def test_nooutlet():
+    p = configparser.ConfigParser()
+    pdu = config_to_power_outlet(p)
+
+    # If we've got no power control then we can't turn the power off:
+    assert pdu.get()
+
+    # No-op
+    pdu.power_on()
+
+    with pytest.raises(Exception):
+        pdu.power_off()