+++ docs/release-notes.md
@@ -5,6 +5,25 @@ These release notes apply to the Stb-tester open-source version. Customers of
 Stb-tester.com Ltd. should refer to the release notes at
 https://stb-tester.com/manual/release-notes instead.
 
+#### v35 (unreleased)
+
+Unreleased
+
+##### Breaking changes since v34
+
+* `stbt power` - Dropped support for PDU types "pdu" and "ipp" as there were
+  several issues with their implementation [#805].  If you were using one of
+  these PDU types please [raise an issue] on GitHub.
+
+[#805]: https://github.com/stb-tester/stb-tester/pull/805
+[raise an issue]: https://github.com/stb-tester/stb-tester/issues/new
+
+##### Major new features
+
+##### Minor additions, bugfixes & improvements
+
+* `stbt power` - Added support for APC7xxx PDUs [#805].
+
 #### v34
 
 14 June 2023.